fix(#S11P12D109-459): 공연상세 조회 쿼리 관련 오류 수정

+++ Colorpl-Backend/src/main/java/com/colorpl/show/repository/ShowDetailRepository.java
@@ -3,7 +3,7 @@ package com.colorpl.show.repository;
 import com.colorpl.show.domain.ShowDetail;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface ShowDetailRepository extends JpaRepository<ShowDetail, Long>,
+public interface ShowDetailRepository extends JpaRepository<ShowDetail, Integer>,
     ShowDetailRepositoryCustom {
 
     boolean existsByApiId(String apiId);
